requirements.txt
# ==================== Core Framework ====================
Flask==3.0.0
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
gunicorn==21.2.0
python-dotenv==1.0.0

# ==================== Database ====================
mysql-connector-python==8.2.0
pymysql==1.1.0
cryptography==41.0.7
SQLAlchemy==2.0.23

# ==================== Cache & Queue ====================
redis==5.0.1
celery==5.3.6

# ==================== Trading & Data ====================
ccxt==4.1.95
requests==2.31.0

# ==================== Scientific Computing ====================
numpy==1.26.2
pandas==2.1.3
scipy==1.11.4

# ==================== Dashboard & Visualization (Phase 1.5) ====================
streamlit==1.29.0
plotly==5.18.0
vectorbt==0.26.2

# ==================== Machine Learning (Optional - Phase 4) ====================
# scikit-learn==1.3.2
# transformers==4.35.2
# torch==2.1.1 --index-url https://download.pytorch.org/whl/cpu

# ==================== Backtesting (已整合到 vectorbt) ====================
# backtrader==1.9.78.123
# matplotlib==3.8.2

# ==================== On-Chain Data (Phase 4 - 可選) ====================
# dune-client==1.3.0
# web3==7.8.0
# eth-typing==5.1.0

# ==================== Quantitative Finance ====================
# cvxpy==1.4.1

# ==================== Notifications ====================
line-bot-sdk==3.5.0

# ==================== Security (Phase 4) ====================
# trezor==0.13.8

# ==================== Testing ====================
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0